+++ AntData/AntData.ORM/SqlQuery/SqlField.cs
@@ -44,7 +44,7 @@ namespace AntData.ORM.SqlQuery
 		public bool             IsUpdatable      { get; set; }
 		public DataType         DataType         { get; set; }
 		public string           DbType           { get; set; }
-		public int?             Length           { get; set; }
+		public long?             Length           { get; set; }
 		public int?             Precision        { get; set; }
 		public int?             Scale            { get; set; }
 		public string           CreateFormat     { get; set; }
